Fix grave interaction and home core protection bugs
- Refactored GraveManager to use String keys for active graves, fixing issues where items wouldn't return or graves wouldn't despawn.
- Updated Armor Stand spawning to use the consumer method, ensuring holograms are invisible and named correctly before appearing to players.
- Fixed ProtectionListener's core block detection to use block coordinates, preventing core blocks from being breakable when set at decimal offsets.
- Improved ShopGUI item removal logic to ensure inventory synchronization across all Bukkit versions.
- Verified compilation and completed code review.

+++ src/main/java/net/wrpnetwork/survivalcore/home/ProtectionListener.java
@@ -44,7 +44,8 @@ public class ProtectionListener implements Listener {
         List<net.wrpnetwork.survivalcore.database.DatabaseManager.HomeData> homes = plugin.getProtectionManager().getWorldHomes(loc.getWorld().getName());
         if (homes == null) return false;
         for (net.wrpnetwork.survivalcore.database.DatabaseManager.HomeData home : homes) {
-            if (home.x() == loc.getBlockX() && home.y() == loc.getBlockY() && home.z() == loc.getBlockZ()) {
+            // Compare block coordinates (int) to ensure it works even if the sethome was slightly offset
+            if ((int)home.x() == loc.getBlockX() && (int)home.y() == loc.getBlockY() && (int)home.z() == loc.getBlockZ()) {
                 return true;
             }
         }
